set up navigation library, migrate navigation component from shared to atl-transit/naviagtion
