+++ libs/features/navigation/src/ui/components/index.ts
@@ -1,0 +1,2 @@
+export * from './mobile-navbar/mobile-navbar.component'
+export * from './sidebar/sidebar.component'
