+++ libs/features/navigation/src/ui/components/mobile-navbar/mobile-navbar.component.html
