+++ libs/features/navigation/src/ui/components/sidebar/sidebar.component.html
