+++ libs/features/navigation/src/ui/views/index.ts
@@ -1,0 +1,1 @@
+export * from './navigation/navigation.component'
