+++ libs/features/navigation/src/ui/views/navigation/navigation.component.scss
