+++ libs/features/navigation/src/ui/views/navigation/navigation.component.ts
@@ -1,8 +1,7 @@
 import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { MobileNavbarComponent } from './mobile-navbar/mobile-navbar.component';
-import { SidebarComponent } from './sidebar/sidebar.component';
+import { MobileNavbarComponent, SidebarComponent } from '../../components';
 
 @Component({
    selector: 'lib-navigation',
@@ -11,6 +10,4 @@ import { SidebarComponent } from './sidebar/sidebar.component';
    templateUrl: './navigation.component.html',
    styleUrls: ['./navigation.component.scss']
 })
-export class NavigationComponent {
-
-}
+export class NavigationComponent {}
