+++ libs/shared/src/content/index.ts
@@ -2,6 +2,4 @@ export * from './button.mock'
 export * from './alert-advisory.mock'
 export * from './dropdown.mock'
 export * from './card-header.mock'
-export * from './page-header.mock'
 export * from './tab.mock'
-export * from './sidebar.mocks'
