+++ libs/shared/src/index.ts
@@ -2,14 +2,12 @@
 export * from './components/advisory-card/advisory-card.component';
 export * from './components/button/button.component';
 export * from './components/card-header/card-header.component';
-export * from './components/navigation/navigation.component';
 export * from './components/hero/hero.component';
 export * from './components/tab/tab.component';
 export * from './components/dropdown/dropdown.component';
 export * from './components/loading-skeleton/loading-skeleton.component';
 export * from './components/inputs/text-input/text-input.component';
 export * from './components/divider/divider.component';
-export * from './components/navigation/mobile-navbar/mobile-navbar.component';
 
 // Exporting Mock, Types, and Interfaces
 export * from './content';
